Accept accessory mode request from Raspberry Pi
The Raspberry Pi sends information over USB to identify itself and then
requests the Samsung Galaxy S4 smartphone to put itself into accessory
mode. Android sends an intent to `UsbTest` to let it know that the Pi
wants to connect, and when `UsbTest` starts, it extracts the model name
of the accessory (“DogBot”) and displays it to the screen.